requirements.txt
Flask-JWT
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
SQLAlchemy==1.2.4
WTForms==2.1
Werkzeug==0.14.1
alembic==0.9.8
argparse==1.2.1
click==6.7
gunicorn==19.7.1
python-dateutil==2.6.1
python-editor==1.0.3

